feat(ui): add cross-app toast store
