+++ packages/ui/src/components/ui/index.d.ts
@@ -1,7 +1,7 @@
 export { Button } from './Button';
 export { Input } from './Input';
 export { Card } from './Card';
-export { Toast, ToastContainer, ConnectedToastContainer } from './Toast';
+export { Toast, ToastContainer, ConnectedToastContainer, useToastStore } from './Toast';
 export { Modal } from './Modal';
 export { Select } from './Select';
 export { Textarea } from './Textarea';
